pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "entities"
version = "0.3.0"
description = "Python SDK for interacting with the Entities Assistant API."
readme = "README.md"
authors = [
    { name = "Francis Neequaye Armah", email = "francis.neequaye@projectdavid.co.uk" }
]
license = { text = "MIT" }
keywords = ["AI", "SDK", "Entities", "LLM", "Assistant"]
dependencies = [
    "httpx>=0.24.1",
    "pydantic>=2.0",
    "python-dotenv>=1.0",
    "aiofiles>=23.2.1",
]

requires-python = ">=3.9"

[project.optional-dependencies]
dev = [
    "black>=23.3",
    "isort>=5.12",
    "pytest>=7.2",
    "mypy>=1.0",
    "build",
    "twine"
]

[project.urls]
"Homepage" = "https://github.com/frankie336/entitites_sdk"
"Documentation" = "https://github.com/frankie336/entitites_sdk/tree/master/docs"
"Changelog" = "https://github.com/your-org/entities-sdk/blob/main/CHANGELOG.md"

[tool.setuptools]
package-dir = { "" = "src" }

[tool.setuptools.packages.find]
where = ["src"]
exclude = ["tests", "docs"]

[tool.setuptools.dynamic]
version = { attr = "_version.__version__" }  # if you're pulling from _version.py

[tool.mypy]
python_version = "3.9"
strict = true

[tool.black]
line-length = 100
target-version = ['py39']

[tool.isort]
profile = "black"
